need to implement hook next
