+++ dialogflow-cx/pytest.ini
@@ -5,3 +5,4 @@ markers =
     hermetic: Marks tests as internally encapsulated
     parameterize: Marks tests as repeated with different parameters
     flaky: Marks tests as flaky for pytest-rerunfailures plugin
+    wip: Marks tests as in-development
