+++ dialogflow-cx/tests/test_validate_form_sample.py
@@ -14,23 +14,12 @@
 
 """Dialogflow CX webhook sample unit tests."""
 
+import os
+import pytest
 import uuid
 
-from google.cloud.dialogflowcx_v3.types import agent as gcdc_agent
-from google.cloud.dialogflowcx_v3 import Agent
-import google.auth
-import pytest
-import os
-from google.cloud.dialogflowcx_v3.services.agents import AgentsClient
-from google.auth import credentials as ga_credentials
-from grpc._channel import _UnaryUnaryMultiCallable
-
-from basic_webhook_sample import BasicWebhookSample
+from validate_form_sample import ValidateFormSample
 from webhook.main import get_webhook_uri
-
-'''
-export TERRAFORM_PROJECT_ID=df-terraform-dev-1113
-'''
 
 
 @pytest.fixture(scope='session')
@@ -55,7 +44,7 @@ def pytest_session_uuid():
 
 @pytest.fixture(scope='session')
 def webhook_sample(project_id, webhook_uri, pytest_session_uuid):
-  sample = BasicWebhookSample(
+  sample = ValidateFormSample(
     agent_display_name = f'Webhook Agent (test session {pytest_session_uuid})',
     project_id=project_id,
     webhook_uri=webhook_uri,
@@ -66,9 +55,9 @@ def webhook_sample(project_id, webhook_uri, pytest_session_uuid):
   del sample
 
 
-@pytest.mark.integration
+@pytest.mark.wip
 @pytest.mark.flaky(max_runs=3, reruns_delay=15)
-@pytest.mark.parametrize("test_case_display_name", BasicWebhookSample.TEST_CASES)
+@pytest.mark.parametrize("test_case_display_name", ValidateFormSample.TEST_CASES)
 def test_indirect(test_case_display_name, webhook_sample):
   test_case_delegator = webhook_sample.test_case_delegators[test_case_display_name]
   if test_case_delegator.expected_exception:
